Day Eighteen Part 1 - done!

+++ src/main/java/advent_of_code_2020/DayEighteen.java
@@ -7,8 +7,6 @@ import java.util.function.BiFunction;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
-import static java.lang.Integer.parseInt;
-
 public class DayEighteen {
 
     private String input;
@@ -18,7 +16,7 @@ public class DayEighteen {
     }
 
 
-    public int evaluate() {
+    public long evaluate() {
         int acc = 0;
 
         Pattern matchedBracketsPattern = Pattern.compile("(.*)\\(([^()]*)\\)(.*)");
@@ -29,7 +27,7 @@ public class DayEighteen {
             matcher = matchedBracketsPattern.matcher(input);
         }
 
-        return Integer.parseInt(squish(input));
+        return Long.parseLong(squish(input));
     }
 
     private String squish(String input) {
@@ -40,8 +38,8 @@ public class DayEighteen {
             if (!matcher.matches()) {
                 throw new IllegalArgumentException("Squish can't match input: " + input);
             }
-            int operand1 = Integer.parseInt(matcher.group(1));
-            int operand2 = Integer.parseInt(matcher.group(3));
+            long operand1 = Long.parseLong(matcher.group(1));
+            long operand2 = Long.parseLong(matcher.group(3));
             switch (matcher.group(2)) {
                 case "+":
                     input = String.valueOf(operand1 + operand2) + matcher.group(4);
